ENH: Expose plugin state via BufExplorer_IsRunning().
I have a custom key mapping that toggles different sidebar panels (Tagbar, Project, BufExplorer, and Undotree) on / off. For that, I need a predicate function that determines whether the plugin is currently active, i.e. shows its window(s). This small public function exposes this information which was previously inaccessible in the script-local variable.

The s:running variable now stores the bufnr of BufExplorer, so :drop does not need to be used to go to its window. (:drop is bad because it modifies the argument list.)